BaseEntity --> com.degloba.domain.support.BaseEntity Entity  --> com.degloba.domain.seedwork.Entity Serializable  --> Long Integer  --> Long MyEntity extends AbstractEntity   -->  MyEntity extends BaseAggregateRoot IRepository --> EntityRepository

+++ degloba-domain/src/test/java/com/degloba/domain/BaseEntityTest.java
@@ -59,21 +59,21 @@ public class BaseEntityTest {
 
     @Test
     public void testExistedIdIsZero() {
-        guangdong.setId(0);
+        guangdong.setId((long) 0);
         assertFalse(guangdong.existed());
     }
 
     @Test
     public void testExistedRepositoryNotFound() {
-        guangdong.setId(3);
-        when(repository.exists(Organization.class, 3)).thenReturn(false);
+        guangdong.setId((long) 3);
+        when(repository.exists(Organization.class,(long) 3)).thenReturn(false);
         assertFalse(guangdong.existed());
     }
 
     @Test
     public void testExistedRepositoryFound() {
-        guangdong.setId(3);
-        when(repository.exists(Company.class, 3)).thenReturn(true);
+        guangdong.setId((long) 3);
+        when(repository.exists(Company.class, (long) 3)).thenReturn(true);
         assertTrue(guangdong.existed());
     }
 
@@ -85,21 +85,21 @@ public class BaseEntityTest {
 
     @Test
     public void testNotExistedIdIsZero() {
-        guangdong.setId(0);
+        guangdong.setId((long) 0);
         assertTrue(guangdong.notExisted());
     }
 
     @Test
     public void testNotExistedRepositoryNotFound() {
-        guangdong.setId(3);
-        when(repository.exists(Company.class, 3)).thenReturn(false);
+        guangdong.setId((long) 3);
+        when(repository.exists(Company.class,(long) 3)).thenReturn(false);
         assertTrue(guangdong.notExisted());
     }
 
     @Test
     public void testNotExistedRepositoryFound() {
-        guangdong.setId(3);
-        when(repository.exists(Company.class, 3)).thenReturn(true);
+        guangdong.setId((long) 3);
+        when(repository.exists(Company.class,(long) 3)).thenReturn(true);
         assertFalse(guangdong.notExisted());
     }
 
@@ -108,15 +108,15 @@ public class BaseEntityTest {
         guangdong.setId(null);
         assertTrue(guangdong.notExisted());
 
-        guangdong.setId(0);
+        guangdong.setId((long) 0);
         assertTrue(guangdong.notExisted());
 
-        guangdong.setId(3);
-        when(repository.exists(Company.class, 3)).thenReturn(false);
+        guangdong.setId((long) 3);
+        when(repository.exists(Company.class,(long) 3)).thenReturn(false);
         assertTrue(guangdong.notExisted());
 
         reset(repository);
-        when(repository.exists(Company.class, 3)).thenReturn(true);
+        when(repository.exists(Company.class,(long) 3)).thenReturn(true);
         assertFalse(guangdong.notExisted());
     }
